supabase_test/requirements.txt
requests
python-dotenv
pandas
openpyxl
schedule
